iRODS filesystem stub and POSIX-to-iRODS factory
We need a factory because using a fresh instantiation of POSIXFilesystem
or iRODSFilesystem will mean creating new objects. This will break the
way the Graph interface works, because it checks for vertex equality
based on reference. One could fix this by using singletons, but it's
conceivable that some filesystems would need specific initialisation
arguments. (I guess these too could be singletons, using a
memoisation-like approach, but then this is getting a bit much!)