Let the database control filesystem load
We know how many tasks are running for each job and we can see which
filesystems are being utilised. The summation by filesystem is the
current filesystem load. If we also include the filesystem's maximum
capacity, we can then further limit the "to do list" view to tasks that
are permissible to be run, given the constraints of the filesystems
involved.

This solves the maximum concurrency problem over multiple, potentially
dependent tasks, purely from the state alone. This frees up the client
to only making the execution capacity available; the theoretical maximum
being the maximum of the pairwise minimums of filesystem tuples used in
a job. For example

  A (max 20) -> B (max 10)  =>  max 10  \    min 5
  B          -> C (max 5)   =>  max 5    }-  avg 6.6...
  C          -> D (max 50)  =>  max 5   /    max 10

i.e., 10 workers would give maximum throughput, at the expense of idle
workers for two stages of a three stage transfer. The mean of 7 workers
(rounded up) would probably be the sweet spot for throughput over
idleness, without having to dynamically manage execution capacity.