Don't auto-start phases on construction
We need to represent uninitialised durations (e.g., for attempts), so
instead of auto-starting phases on construction, we set their state to
that of what's persisted, if anything. This makes the logic a bit more
obvious and allows checking for phase starts, but it means that phases
will need to be manually started in application code.