Allow attempts to have a null start time
We need to create a sentinel attempt record to remove the task from the
todo list view. In practice, the start time will be set almost
immediately after creating that record, but not at the same time. This
will allow us to lock the tables controlling the todo list for as short
a time as possible.