Cache the results from baton
We currently call baton FOUR times to determine the size and checksum of
each file, when we only need to call it once and cache the result. This
is particularly significant when using our containerised version of
baton, which has a non-trivial start-up time. Using the stdlib lru_cache
decorator will sort this out.